requirements.txt
fastapi==0.115.0
mangum==0.19.0
psycopg2-binary==2.9.9
